+++ source/extensions/lightspeed.event.capture_persp_to_persp/lightspeed/event/capture_persp_to_persp/tests/unit/test_core.py
@@ -20,7 +20,7 @@ from lightspeed.event.capture_persp_to_persp.core import CopyCapturePerspToPersp
 from lightspeed.layer_manager.core import LayerManagerCore as _LayerManagerCore
 from lightspeed.layer_manager.core import LayerType as _LayerType
 from lightspeed.trex.capture.core.shared import Setup as _CaptureCoreSetup
-from lightspeed.trex.viewports.shared.widget import SetupUI as ViewportUI
+from lightspeed.trex.viewports.shared.widget import create_instance as _create_viewport_instance
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
 from omni.kit.test.async_unittest import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, open_stage
@@ -34,7 +34,9 @@ _CONTEXT_NAME = ""
 @contextlib.asynccontextmanager
 async def make_temp_directory(context):
     temp_dir = tempfile.TemporaryDirectory()  # noqa PLR1732
-    viewport = ViewportUI(_CONTEXT_NAME)  # create the viewport object because the viewport generate the persp camera
+    viewport = _create_viewport_instance(
+        _CONTEXT_NAME
+    )  # create the viewport object because the viewport generate the persp camera
     try:
         yield temp_dir
     finally: