+++ source/extensions/lightspeed.trex.viewports.shared.widget/lightspeed/trex/viewports/shared/widget/__init__.py
@@ -1,2 +1,12 @@
-from .extension import TrexViewportSharedExtension
-from .setup_ui import SetupUI  # noqa F401
+"""
+* Copyright (c) 2022, NVIDIA CORPORATION.  All rights reserved.
+*
+* NVIDIA CORPORATION and its licensors retain all intellectual property
+* and proprietary rights in and to this software, related documentation
+* and any modifications thereto.  Any use, reproduction, disclosure or
+* distribution of this software and related documentation without an express
+* license agreement from NVIDIA CORPORATION is strictly prohibited.
+"""
+__all__ = ["create_instance", "get_instance", "get_instances"]
+
+from .extension import TrexViewportSharedExtension, create_instance, get_instance, get_instances