+++ source/extensions/lightspeed.event.capture_persp_to_persp/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1] - 2023-12-11
+### Changed
+- Minor change for an API rename
+
 ## [1.1.0] - 2023-10-10
 ### Changed
 - cleanup