+++ source/extensions/lightspeed.trex.layout.shared.mass_ingestion/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.1] - 2023-12-11
+### Changed
+- Minor change for API rename (ViewportUI)
+
 ## [1.0.0] - 2022-06-09
 ### Added
 - Created