+++ source/extensions/lightspeed.trex.layout.stagecraft/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.0.2] - 2023-12-11
+### Changed
+- Minor change for API rename (ViewportUI)
+
 ## [1.0.1] - 2023-12-07
 ### Changed
 - Changed "Layout" to "Modding"